+++ server/src/main/java/com/linbit/linstor/storage/interfaces/categories/resource/AbsRscLayerObject.java
@@ -51,10 +51,6 @@ public interface AbsRscLayerObject<RSC extends AbsResource<RSC>>
 
     void remove(AccessContext accCtx, VolumeNumber vlmNrRef) throws AccessDeniedException, DatabaseException;
 
-    boolean checkFileSystem();
-
-    void disableCheckFileSystem();
-
     boolean exists();
 
     void setIsSuspended(boolean suspend) throws DatabaseException;