+++ server/src/main/java/com/linbit/linstor/storage/interfaces/categories/resource/VlmProviderObject.java
@@ -165,4 +165,9 @@ public interface VlmProviderObject<RSC extends AbsResource<RSC>> extends LayerOb
     {
         return getVolumeKey().compareTo(oRef.getVolumeKey());
     }
+
+    boolean checkFileSystem();
+
+    void disableCheckFileSystem();
+
 }